refactor(ui): estructura de carpetas
- Se corrigio la estructura de carpetas del proyecto.
