+++ src/app/shared/components/loader/loader.component.html
