+++ src/app/shared/components/loader/loader.component.ts
